COV-15585: code cleanup

+++ com.openexchange.admin.plugin.hosting/src/com/openexchange/admin/storage/mysqlStorage/OXContextMySQLStorage.java
@@ -2705,7 +2705,7 @@ public class OXContextMySQLStorage extends OXContextSQLStorage {
             if (null != cacheService) {
                 try {
                     final Cache jcs = cacheService.getCache("CapabilitiesContext");
-                    final Serializable key = Integer.valueOf(ctx.getId().intValue());
+                    final Serializable key = ctx.getId();
                     jcs.remove(key);
                 } catch (final Exception e) {
                     LOG.error("", e);
